class SyncTracker externalized
